Fix fast8 tox target
This is a major timesaver for repos the size of nova. This broke
recently due to changes in flake8 itself. This removes some
needless complexity to make it work again. It also removes the
suggestion to use pre-commit which has nothing to do with this
target, which also stings more when pre-commit is breaking things
which is why you're using this in the first place.

Change-Id: Ieb150bf0931ad8031ca83bae1f206075a9f505e2
Signed-off-by: Dan Smith <dansmith@redhat.com>