nova-conductor puts instance in error state
Nova-conductor puts instance in error if an unknown exception is raised
in the _build_live_migrate_task during the live-migration. [1]
The exception comes from _call_livem_checks_on_host and we can see raise
exception.MigrationPreCheckError if we face to
messaging.MessagingTimeout exception for example. [2]
The function check_can_live_migrate_destination does a check also on source
host with check_can_live_migrate_source [3] and this check can also
return exceptions like MessagingTimeout and this one is not caught properly
because it's a remote "Remote error: MessagingTimeout" due to dest host try to
contact source host and this source host not reply.

[1] https://github.com/openstack/nova/blob/master/nova/conductor/manager.py#L523
[2] https://github.com/openstack/nova/blob/master/nova/conductor/tasks/live_migrate.py#L381
[3] https://github.com/openstack/nova/blob/master/nova/compute/manager.py#L9090

Closes-Bug: #2044235
Change-Id: Ie1f96fee743c235ab35113a9ad1549a67b975839
Signed-off-by: Julien Le Jeune <julien.le-jeune@ovhcloud.com>