Move monkey_patch from init to the entrypoints
This move is needed so that we can define a per service default for
monkey patching.

And yes, the single line with both noqa and autopep8 decorators are
needed to convince autopep8 that this code is OK to be at the start of
the file.

After moving the monkey_patching earlier in the wsgi entrypoint I needed
to move the functional test monkey_patching call earlier too to keep the
early enough for the test where the wsgi entry point is not directly
imported

Change-Id: Idedd2a440adc1cde1e8ffe6636854d5a891e66d2
Signed-off-by: Balazs Gibizer <gibi@redhat.com>