Fix test_simple_tenant_usage test
API policy test_simple_tenant_usage test does
not send the start and end time in request's query
string. In that case, API set the current time to
both start and end times. So there is a chance that
both start and end times can be the same, and Nova
raises an error:
- https://github.com/openstack/nova/blob/9e5ad07aeeb9f14eba37e2cdea9377e7af48ef88/nova/api/openstack/compute/simple_tenant_usage.py#L258

Closes-Bug: 2130703

Change-Id: Ib47890087110d460504df64aeed5206ded2e70b0
Signed-off-by: Ghanshyam Maan <gmaan@ghanshyammann.com>