Fix usage count when user-scoped quota is set
Currently there is an assumption in the code that if a user-scoped
resource usage count is present, user quota has been exceeded.

This is not necessarily true however because when a quota is counted
across both a project and a user, it is possible for the project quota
to be exceeded while the user quota is not.

For example, if the project ram quota is 4000 and user A has a
user-scoped ram quota of 4000, with user A using 512 ram and user B
using 2048 ram (2560 total), a create request by user A for 2048 more
ram should be rejected (project total would be 4608 > 4000). Before
this fix, the code assumed user quota was exceeded since a user-scoped
count existed, and reported user A's usage (512) instead of project
usage (2560). This caused incorrect headroom to be computed, leading
to infinite recursion and a 500 error.

This adds the scope 'project' or 'user' to the OverQuota exception
when it is raised so the correct usage can be reported.

This also adds more detail to existing code comments to aid in future
debugging. The difference between "project quota", "project user
quota", and "user quota" in the code can be difficult to remember over
time.

Functional tests are added to assert that the API does not allow a user
quota to be set larger than the project quota, nor a project quota to
be set smaller than an existing user quota in the project.

Closes-Bug: #2131272

Change-Id: I75577c62392bfa802bbaf7a4eed1859d07d41418
Signed-off-by: melanie witt <melwittt@gmail.com>