[Trivial]: Remove unused RBAC default alias
Nova policy defaults use the rule alias from base class.
This helps to achieve the consistency and avoid redefining
the common rules.

This commit keeps the usable alias in base class and remove
the unused one: PROJECT_MANAGER, PROJECT_MEMBER, and
PROJECT_MEMBER (even they should not be used as they do not
let admin to access the APIs).

Change-Id: I5dcb15306f35e2fc00c0780d3db74bf4453b1cab
Signed-off-by: Ghanshyam <gmaan@ghanshyammann.com>