Make QEMU_IMG_LIMITS process limits configurable
Currently the CPU time and address space process limits for qemu-img
are hard-coded to 30 seconds and 1G respectively.

With more recent versions of Ceph in upstream CI, we have experienced
test failures that suggest 1G is no longer large enough for encrypted
RBD images. In the failures the following error is raised:

  nova.exception.InvalidDiskInfo: Disk info file is invalid: qemu-img
    failed to execute on
    rbd:volumes/volume-c83c9b7f-0f38-4bb8-a40a-300a66080d21:id=cinder :
    Unexpected error while running command.
  Command: /opt/stack/data/venv/bin/python3.12 -m
    oslo_concurrency.prlimit --as=1073741824 --cpu=30 -- env LC_ALL=C
    LANG=C qemu-img info
    rbd:volumes/volume-c83c9b7f-0f38-4bb8-a40a-300a66080d21:id=cinder
    --force-share --output=json
  Exit code: -6
  Stdout: ''
  Stderr: 'failed to allocate memory for stack: Cannot allocate memory\n'

This adds config options ``images_cpu_time_limit`` and
``images_address_space_limit`` to the ``[libvirt]`` section to allow
for tuning of the qemu-img process limits, similar to how Cinder and
Ironic make qemu-img process limits configurable.

Closes-Bug: #2116852

Change-Id: I10e53de27b063b1e514e04066d0eb56a86188e9a
Signed-off-by: melanie witt <melwittt@gmail.com>