Remove eventlet from CacheConcurrencyTestCase
This patch generalize two unit tests from the libvirt driver coverage
that used direct eventlet.spawn calls to test concurrent behavior of the
ImageBackend. Both tests now using util.spawn to make the threads green
or native depending on the environment. Also we needed to adjust some
logic in the tests that relied on eventlet's run to completion semantic.
In one place the cooperative yield was replaced with a real sleep to let
the native thread actually finish its job. In another place the test
instead of asserting the state of an Event, needs to wait for that event
to be set.

Signed-off-by: Balazs Gibizer <gibi@redhat.com>
Change-Id: I4d3e34d6b57a49cb8cb5fb0557e4b4b4e281fd80